chore: spin off fluentbogus/relation into its own project
