+++ NineteenSevenFour.Testing.Core.UnitTest/Model/AddressModel.cs
@@ -1,4 +1,4 @@
-namespace NineteenSevenFour.Testing.FluentBogus.UnitTest.Model;
+namespace NineteenSevenFour.Testing.Core.UnitTest.Model;
 
 #pragma warning disable CS8618 // Non-nullable field must contain a non-null value when exiting constructor. Consider declaring as nullable.
 public class AddressModel
