+++ NineteenSevenFour.Testing.Core.UnitTest/Model/AddressType.cs
@@ -1,0 +1,8 @@
+namespace NineteenSevenFour.Testing.Core.UnitTest.Model;
+
+public enum AddressType
+{
+  Personal,
+  Business,
+  Other
+}
