+++ NineteenSevenFour.Testing.Core.UnitTest/Model/PersonType.cs
@@ -1,0 +1,8 @@
+namespace NineteenSevenFour.Testing.Core.UnitTest.Model;
+
+public enum PersonType
+{
+  Adult,
+  Child,
+  Infant
+}
