+++ NineteenSevenFour.Testing.Core.UnitTest/NineteenSevenFour.Testing.Core.UnitTest.csproj
@@ -1,0 +1,15 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFramework>net6.0</TargetFramework>
+    <ImplicitUsings>enable</ImplicitUsings>
+    <Nullable>enable</Nullable>
+    <IsPackable>false</IsPackable>
+    <ApplicationIcon>..\NineteenFourSeven_logo.ico</ApplicationIcon>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Content Include="..\NineteenFourSeven_logo.ico" />
+  </ItemGroup>
+
+</Project>
