+++ NineteenSevenFour.Testing.FluentBogus.AutoMapper.UnitTest/Usings.cs
@@ -1,3 +1,2 @@
 global using AutoBogus;
-global using Xunit;
 global using NineteenSevenFour.Testing.Core.UnitTest.Model;
