+++ NineteenSevenFour.Testing.FluentBogus.Relation.UnitTest/UnitTest1.cs
@@ -1,0 +1,11 @@
+namespace NineteenSevenFour.Testing.FluentBogus.Relation.UnitTest
+{
+  public class UnitTest1
+  {
+    [Fact]
+    public void Test1()
+    {
+
+    }
+  }
+}
