--- NineteenSevenFour.Testing.FluentBogus.UnitTest/Model/AddressType.cs
@@ -1,8 +1,0 @@
-﻿namespace NineteenSevenFour.Testing.FluentBogus.UnitTest.Model;
-
-public enum AddressType
-{
-  Personal,
-  Business,
-  Other
-}
