--- NineteenSevenFour.Testing.FluentBogus.UnitTest/Model/PersonType.cs
@@ -1,8 +1,0 @@
-﻿namespace NineteenSevenFour.Testing.FluentBogus.UnitTest.Model;
-
-public enum PersonType
-{
-  Adult,
-  Child,
-  Infant
-}
