+++ NineteenSevenFour.Testing.FluentBogus/Interface/IFluentBogusBuilder[TFaker].cs
@@ -4,16 +4,6 @@ using System.Linq.Expressions;
 
 namespace NineteenSevenFour.Testing.FluentBogus.Interface;
 
-
-public interface IFluentBogusBuilder<TEntity>
-    where TEntity : class
-{
-  IFluentBogusBuilder<AutoFaker<TEntity>, TEntity> WithDefault(params object?[]? args);
-
-  IFluentBogusBuilder<TFaker, TEntity> With<TFaker>(params object?[]? args)
-      where TFaker : AutoFaker<TEntity>, new();
-}
-
 public interface IFluentBogusBuilder<TFaker, TEntity>
     where TFaker : AutoFaker<TEntity>, new()
     where TEntity : class
