+++ NuGet.Config
@@ -1,0 +1,14 @@
+<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+	<config>
+		<add key="defaultPushSource" value="https://nuget.pkg.github.com/nineteensevenfour/" />
+	</config>
+
+	<packageSources>
+		<clear />
+		<!-- ensure only the sources defined below are used -->
+		<add key="Nuget" value="https://api.nuget.org/v3/index.json" />
+		<add key="Github" value="https://nuget.pkg.github.com/nineteensevenfour/index.json" />
+	</packageSources>
+
+</configuration>
